refactor: simplify exception handling and rename clothing items use case
