--- src/main/kotlin/data/repository/mapper/ConfigMapper.kt
@@ -1,4 +1,0 @@
-package org.example.data.repository.mapper
-
-class ConfigMapper {
-}
