--- src/main/kotlin/logic/entity/WeatherWithUvIndex.kt
@@ -1,10 +1,0 @@
-package logic.entity
-
-
-data class WeatherWithUv(
-    val temperature: Double,
-    val precipitation: Double,
-    val windSpeed: Double,
-    val humidity: Double,
-    val uvIndex: Double
-)
