+++ src/main/kotlin/logic/exceptions/ApiException.kt
@@ -1,6 +1,5 @@
 package logic.exceptions
 
-sealed class ApiException(message: String, cause: Throwable? = null) : Exception(message, cause) {
-    class NetworkError(message: String, cause: Throwable? = null) : ApiException(message, cause)
-    class DataError(message: String, cause: Throwable? = null) : ApiException(message, cause)
-}
+
+class NetworkError(message: String) : Exception(message)
+class DataError(message: String) : Exception(message)
