+++ src/main/kotlin/logic/usecase/SuggestClothingUseCase.kt
@@ -1,4 +1,4 @@
 package org.example.logic.usecase
 
-interface SuggestClothingUseCase {
+class SuggestClothingUseCase {
 }
