--- src/main/kotlin/logic/usecase/SuggestClothingUseCaseImpl.kt
@@ -1,4 +1,0 @@
-package org.example.logic.usecase
-
-class SuggestClothingUseCaseImpl {
-}
